backend_gcp_cloud_run/requirements.txt
Flask
pytest==8.3.2
ffmpeg-python
pydub
google-cloud-aiplatform
moviepy
gunicorn
azure-cognitiveservices-speech
elevenlabs
python-dotenv
python-multipart
uvicorn
google-cloud-storage
starlette
flask-cors